+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 3.2.3
 
 * Spell Effect: Improvements to Chromatic Orb (Thanks to @Elwin#1410)
+* Spell Effect: Color Spray
 
 # 3.2.2